Now reviews can succesfully be added
When the user first clicks the game, the reviews can be seen on the VBox at the right. Moreover, when the user adds a review by clicking the "Add review" button, the review is succesfully added to both the database and to the VBox. However, a current problem is that the games are added over and over again after each click, and the review of a game is sometimes added to another game's VBox.